Create WHO-App-Relicense.md
Hi @amourakora,
You're the last significant contributor I needed to track down! Please can you comment
"I Agree" on this GitHub issue to support relicensing the WHO App under the Apache 2.0
License. 53 people and 98% of contributors have already said yes. You're the last 2%!!

https://github.com/WorldHealthOrganization/app/issues/1645

We're getting close to launching the WHO App but I need your help. The Apache 2.0
License is a requirement for the Linux Foundation for Public Health to potentially
host the WHO App project. See FAQ for more details.

Thanks and please do it quickly as it's a big project and it takes a lot of time
to chase down everyone. Thanks.

Bruno